requirements.txt
fastapi
uvicorn
python-multipart
inference-sdk
google-generativeai
python-dotenv
